reportとcommentに「編集・削除時の本人確認」を追加

+++ app/controllers/comments_controller.rb
@@ -3,7 +3,11 @@
 class CommentsController < ApplicationController
   def destroy
     comment = Comment.find(params[:id])
-    comment.destroy
-    redirect_to @commentable, notice: 'Comment was successfully destroyed.'
+    if current_user_id?(comment.user_id)
+      comment.destroy
+      redirect_to @commentable, notice: 'Comment was successfully destroyed.'
+    else
+      redirect_to root_path
+    end
   end
 end
